pyproject.toml
[tool.poetry]
name = "bank-challenge-ml-project"
version = "0.1.0"
description = ""
authors = ["Artur Tiburski Vaz Costa <artur_vazcosta@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
scikit-learn = "^1.5.0"
xgboost = "^2.0.3"
pyspark = "3.5.0"
google-cloud-bigquery = "3.14.1"
google-cloud-bigquery-storage = "2.24.0"
pandas-gbq = "0.20.0"
pandas = "^2.2.2"
numpy = "^1.26.4"
seaborn = "^0.13.2"
matplotlib = "^3.9.0"
shap = "^0.45.1"
scipy = "^1.13.1"
lightgbm = "^4.3.0"
scikit-optimize = "^0.10.2"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
pydantic = "^2.7.4"
streamlit = "^1.36.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"